langchain-langgraph-edenmarco/section-6-rag-embeddings-vector-dbs/pyproject.toml
[project]
name = "section-6-rag-embeddings-vector-dbs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "faiss-cpu>=1.12.0",
    "isort>=7.0.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langchain-pinecone>=0.2.13",
    "langchainhub>=0.1.21",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
]
